+++ src/concurrency.c
@@ -1,4 +1,5 @@
-#include "misc.h"
+#include "assertion.h"
+#include "utility.h"
 #include "logging.h"
 #include "concurrency.h"
