+++ src/evaluate.h
@@ -10,7 +10,6 @@
 
 #include <launchdarkly/json.h>
 #include <launchdarkly/variations.h>
-#include <launchdarkly/store.h>
 
 #include "store.h"
