+++ src/logging.c
@@ -3,7 +3,8 @@
 
 #include <launchdarkly/api.h>
 
-#include "misc.h"
+#include "assertion.h"
+#include "utility.h"
 
 static LDLogLevel sdkloggerlevel = LD_LOG_INFO;
 static void (*sdklogger)(const LDLogLevel level, const char *const text) = NULL;
