+++ src/operator.c
@@ -6,8 +6,9 @@
 
 #include <launchdarkly/api.h>
 
+#include "assertion.h"
 #include "operators.h"
-#include "misc.h"
+#include "utility.h"
 
 #define CHECKSTRING(uvalue, cvalue) \
     if (LDJSONGetType(uvalue) != LDText || \
