+++ tests/test-allflags.c
@@ -1,9 +1,10 @@
 #include <launchdarkly/api.h>
 
+#include "assertion.h"
 #include "client.h"
 #include "config.h"
 #include "evaluate.h"
-#include "misc.h"
+#include "utility.h"
 #include "util-flags.h"
 #include "store.h"
